Add toast err msg system, kind of; None node name
if the user has not entered anything for the node name then
it will be passed as None to the callback. added code to acct for
this and throw a UIError.

speaking of UIErrors... we should ideally handle things so that
they trigger toast messages! i am sure there is some fancy way to
modify the error class to trigger the callback but hmmm i am not
sure how to manually do that in dash