-/+ btns for cc size rank
there is some jank about the number input and passing None to
our funcs (see eg
https://stackoverflow.blog/2022/12/26/why-the-number-input-is-the-worst-input/)
and i am not 100% happy with how too-large numbers are treated
as invalid, turned into None from my code's perspective, and thus
turned into 1. would be nice to fix that. but this is fine for now