random node/edge colors! & edge arrowheads
lots of good stuff here.

there is some slight jank where, like, if you have a node or edge
selected and then apply a diff colorization method, then the graph
might "flash" with the old colorization when you next interact
with it. this might be a side effect of us updating the stylesheet
in a hacky way rather than having cy be persistent

but fine for now

closes #53

other thing that would be nice: assigning similar or identical
random colors to a node and its twin if present