Consistent random colors for nodes & kinda edges
+/- nodes, split nodes, and fake edges all use same colors.

this was actually a v satisfying solution

not applied yet for ordinary (not fake) edges. hm. if the graph
is not a multigraph then it is straightforward to associate an
edge with its RC, but if the graph IS a multigraph then that is a
big pain - like in LJA we have to look at the fasta file of seqs
right? ew

perhaps we can do something simple - check if an edge does not
have a "sibling" parallel to it, and if so, just set their rand
indices identical. and otherwise do nothing? but hmmmmmm then
you get jank with like, people assuming that this will also apply
to multigraphs and then ahsdouhsodifhasodif