Remove "Apply" button for node/edge coloring
Instead, just have changes to the radio buttons' values trigger the
callback directly.

Besides being snappy, this has the advantage of avoiding weird
situations where someone changes a radio button but then hits
"draw" instead of "apply" -- in such cases the original settings
would persist since "apply" had not been hit, which is maybe
a bit confusing / unintuitive.

We COULD keep the apply button and instead just, like, have the
cy stylesheet be returned by draw() and idk have the apply button
also talk to draw(), but that sounds confusing and gross and
probably not the idiomatic way to do this.

if performance is a bottleneck then perhaps we can add back the
apply button later but i think this will be ok