include patterns in viz!
a couple of other good changes here, too, including setting
a better name for Patterns (not sure i want to actually keep this
around tho) and refactoring to_cyjs() to be a component by component
thing.

this makes to_dot() look kind of ugly by comparison lmao. not that
we are even using that??? i guess it will be helpful for debugging
but like... we don't even use it when we call pygraphviz normally lol