exclude other new config files from code coverage
probably doesn't matter all that much. if we start adding like
any remotely fancy logic to these files then I GUESS we could
add them to the coverage but probs nbd

same with to include or not to include test code. including it
inflates coverage to ~79% but this is kind of unrealistic and
probs only really useful for diagnosing ignored tests atm imo