ok actually just consistently install pgv w/ conda
It is good to know that we don't strictly NEED to install pygraphviz
before installing mgsc (like, we can just install graphviz first
and then install mgsc, with the understanding that pygraphviz being
listed as a dependency will get it to be installed).

however, if we are going to tell people in the README to install
pygraphviz using conda (since this seems a bit safer / less prone
to breaking), thennnnnn we should probably test things that way
also. I mean i guess we COULD test both installation methods but
i don't think this is important enough to care about right now