Add edge IDs to old chr15/21 LJA graphs
this allows these graphs to be loaded in metagenomescope!

It looks like the chr15_subgraph test was already broken -- looking
at my notes from a few years ago i see the bug documented there in
the current visualization of that graph. so, that test will fail,
and now we can focus on fixing this bug (a node is not being
merged back after splitting)