Add back a fit-graph-to-screen button
this was kind of painful -- Dash-Cytoscape doesn't give easy access
(as far as I can tell) to the underlying Cytoscape.js instance,
so now we need to do a bit more thinking before just calling
cy.fit() or something.

Thankfully, there is a workaround for getting the underlying
Cytoscape.js instance after it has been created (shoutouts to
https://stackoverflow.com/a/52603597) that appears to be relatively
stable. This lets us essentially get back to the days of just calling
cy.fit(), etc. directly.

This workaround may break in the future, but we can cross that bridge
when we get to it. For now I think this should be fine.