use dash bootstrap components; fancy radio btns
precursor to image export stuff -- i wanted fancy radio buttons
for that and figured oh, might as well apply to the colorization
stuff also.

thankfully it looks like we can use dbc a la carte -- so no need
to go ahead with updating EVERYTHING. eventually it would be nice
to adjust the modal dialogs, etc to use dbc (since this reduces
boilerplate a lot), but for now it is fine.