More intelligent ETFE trimming: close #267
This fixes the chr15 problem with a leftover unsplit node! It seems
like what was happening was -- when we created this goofy chain
(ID 107) containing a chain and then this one rightmost cyclic chain
that should've been included in the first place, this messed with
the 262-L node somehow and caused it to stick around. Something about
edge rerouting / reverting in the decomposed graph. yuck.

I am not convinced that the evil is entirely defeated (there may be
other weird problems that crop up), but the tests should at least
pass now.