Use distinct shapes for split nodes
:)

not sure yet how the semicircle stuff will work for large graphs...
we are using 23 points for each semicircle's polygon, if i'm counting
correctly?

compared to just 5 for each "house" / "invhouse" polygon.

it should be easy to decrease the amount of stuff in `xpts` in
config.py in order to decrease the amount of points to be drawn