more accurate split pentagon shapes; add util func
You can now see clearly that yeah, the left split + the right split
can be combined into a full node.

An issue with this is that the triangles are arguably too small...
We can adjust the dimensions (this is easier to do now that there
is a nice util func for creating shape-polygon-points strings!)
or we could just wait and see how things look when we start
scaling nodes by lengths again.

and a third solution would be adjusting this so that the triangles
also include a bit of the "rectangle" as well? but then this messes
with visual identification of split nodes...