Add searching for node(s) & multi-node cc select
Lot of changes here. This implements things and works well but
there is a lot of room to be done on polishing up the toast messages,
tidying up the code, etc.

Also it might be worth moving the JS to its own file instead of
being in a docstring. Writing complicated JS without syntax
highlighting is a pain