Put controls <div> right border inside scrollbar
seems fine -- looks a lot better than previous

minor issues:

- when you zoom in super far, to the point where the control panel
has a horizontal scrollbar appear, the border is not shown until you
like manually scroll the panel to the right. This does not happen
when the border and scrollbar are on the same element. Not really
a big deal imo

- also when you zoom in super far, the dropdown menu that shows up
when you click on the button is cut off by the right edge of the
control panel. In theory
https://github.com/twbs/bootstrap/issues/24251
or something should fix this, but I haven't had luck with setting
`data-bs-boundary: viewport;` on the dropdown yet -- will need to
test. also i am not sure if this was already a problem before this
commit, need to check