Make pressing "Enter" in the sr / nn <input>s draw
Nice convenience thing. Closes #250.

There is the lingering worry that "what if the user changes to
another cc selection method but somehow still retains focus on
the hidden <input> -- would Enter still redraw the graph then?"
but that should never happen because changing the cc selection
method should require clicking on the dropdown for that and thus
losing focus. also that is not even a bug really (it will just draw
the graph using whatever method is specified in the dropdown).