style agg cc rects in treemap with purple gradient
This mostly supplants #278, I think - it gets the point across
(there is something unique about these rectangles) while avoiding
just using the same style for all of them, and avoiding the weird
bug with disappearing tiny rectangles that seems to be induced by
nesting things. yay!

the code that computes the gradient is kinda messy but whatever
lets focus on the actual important stuff for now (plus we might
change this eventually so no need to calcify it quite yet imo)